feat: sistema de autenticación
Se implementa sistema completo de autenticación con manejo de distintos proveedores

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getServerUserSession } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma';
 import { NextRequest, NextResponse } from 'next/server';
 import * as yup from 'yup';
@@ -39,9 +40,15 @@ const postSchema = yup.object( {
 
 export async function POST ( request: NextRequest ) {
 
+    const user = await getServerUserSession();
+
+    if ( !user ) {
+        return NextResponse.json( 'No autorizado', { status: 401 } );
+    }
+
     try {
         const { description } = await postSchema.validate( await request.json() );
-        const newTodo = await prisma.todo.create( { data: { description } } );
+        const newTodo = await prisma.todo.create( { data: { description, userId: user.id! } } );
 
         return NextResponse.json( {
             ok: true,
@@ -57,9 +64,14 @@ export async function POST ( request: NextRequest ) {
 
 
 export async function DELETE ( request: NextRequest ) {
+    const user = await getServerUserSession();
+
+    if ( !user ) {
+        return NextResponse.json( 'No autorizado', { status: 401 } );
+    }
 
     try {
-        await prisma.todo.deleteMany( { where: { complete: true } } );
+        await prisma.todo.deleteMany( { where: { complete: true, userId: user.id } } );
 
         return NextResponse.json( {
             ok: true,
